requirements.txt
attrs==25.4.0
cffi==2.0.0
cryptography==46.0.3
freezegun==1.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
pytest==8.4.2
referencing==0.37.0
rpds-py==0.28.0
typing_extensions==4.15.0
PySocks==1.7.1
